Moved tests across classes
