+++ tests/PrivateChannelTest.php
@@ -3,6 +3,7 @@
 namespace BeyondCode\LaravelWebSockets\Test;
 
 use BeyondCode\LaravelWebSockets\Server\Exceptions\InvalidSignature;
+use Carbon\Carbon;
 use Ratchet\ConnectionInterface;
 
 class PrivateChannelTest extends TestCase
@@ -152,5 +153,114 @@ class PrivateChannelTest extends TestCase
                     );
                 }
             });
+    }
+
+    public function test_not_ponged_connections_do_get_removed_for_private_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $activeConnection = $this->newPrivateConnection('private-channel');
+        $obsoleteConnection = $this->newPrivateConnection('private-channel');
+
+        // The active connection just pinged, it should not be closed.
+        $this->channelManager->addConnectionToSet($activeConnection, Carbon::now());
+
+        // Make the connection look like it was lost 1 day ago.
+        $this->channelManager->addConnectionToSet($obsoleteConnection, Carbon::now()->subDays(1));
+
+        $this->channelManager
+            ->getGlobalConnectionsCount('1234', 'private-channel')
+            ->then(function ($count) {
+                $this->assertEquals(2, $count);
+            });
+
+        $this->channelManager
+            ->getConnectionsFromSet(0, Carbon::now()->subMinutes(2)->format('U'))
+            ->then(function ($expiredConnections) {
+                $this->assertCount(1, $expiredConnections);
+            });
+
+        $this->channelManager->removeObsoleteConnections();
+
+        $this->channelManager
+            ->getGlobalConnectionsCount('1234', 'private-channel')
+            ->then(function ($count) {
+                $this->assertEquals(1, $count);
+            });
+
+        $this->channelManager
+            ->getConnectionsFromSet(0, Carbon::now()->subMinutes(2)->format('U'))
+            ->then(function ($expiredConnections) {
+                $this->assertCount(0, $expiredConnections);
+            });
+    }
+
+    public function test_events_are_processed_by_on_message_on_private_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $connection = $this->newPrivateConnection('private-channel');
+
+        $message = new Mocks\SignedMessage([
+            'appId' => '1234',
+            'serverId' => 'different_server_id',
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'private-channel',
+                'test' => 'yes',
+            ],
+        ], $connection, 'private-channel');
+
+        $this->channelManager->onMessage(
+            $this->channelManager->getRedisKey('1234', 'private-channel'),
+            $message->getPayload()
+        );
+
+        // The message does not contain appId and serverId anymore.
+        $message = new Mocks\SignedMessage([
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'private-channel',
+                'test' => 'yes',
+            ],
+        ], $connection, 'private-channel');
+
+        $connection->assertSentEvent('some-event', $message->getPayloadAsArray());
+    }
+
+    public function test_events_get_replicated_across_connections_for_private_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $connection = $this->newPrivateConnection('private-channel');
+        $receiver = $this->newPrivateConnection('private-channel');
+
+        $message = new Mocks\SignedMessage([
+            'appId' => '1234',
+            'serverId' => $this->channelManager->getServerId(),
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'private-channel',
+                'test' => 'yes',
+            ],
+            'socketId' => $connection->socketId,
+        ], $connection, 'private-channel');
+
+        $channel = $this->channelManager->find('1234', 'private-channel');
+
+        $channel->broadcastToEveryoneExcept(
+            $message->getPayloadAsObject(), $connection->socketId, '1234', true
+        );
+
+        $receiver->assertSentEvent('some-event', $message->getPayloadAsArray());
+
+        $this->getSubscribeClient()
+            ->assertNothingDispatched();
+
+        $this->getPublishClient()
+            ->assertCalledWithArgs('publish', [
+                $this->channelManager->getRedisKey('1234', 'private-channel'),
+                $message->getPayload(),
+            ]);
     }
 }
