+++ tests/PublicChannelTest.php
@@ -2,6 +2,7 @@
 
 namespace BeyondCode\LaravelWebSockets\Test;
 
+use Carbon\Carbon;
 use Ratchet\ConnectionInterface;
 
 class PublicChannelTest extends TestCase
@@ -133,5 +134,114 @@ class PublicChannelTest extends TestCase
                     );
                 }
             });
+    }
+
+    public function test_not_ponged_connections_do_get_removed_for_public_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $activeConnection = $this->newActiveConnection(['public-channel']);
+        $obsoleteConnection = $this->newActiveConnection(['public-channel']);
+
+        // The active connection just pinged, it should not be closed.
+        $this->channelManager->addConnectionToSet($activeConnection, Carbon::now());
+
+        // Make the connection look like it was lost 1 day ago.
+        $this->channelManager->addConnectionToSet($obsoleteConnection, Carbon::now()->subDays(1));
+
+        $this->channelManager
+            ->getGlobalConnectionsCount('1234', 'public-channel')
+            ->then(function ($count) {
+                $this->assertEquals(2, $count);
+            });
+
+        $this->channelManager
+            ->getConnectionsFromSet(0, Carbon::now()->subMinutes(2)->format('U'))
+            ->then(function ($expiredConnections) {
+                $this->assertCount(1, $expiredConnections);
+            });
+
+        $this->channelManager->removeObsoleteConnections();
+
+        $this->channelManager
+            ->getGlobalConnectionsCount('1234', 'public-channel')
+            ->then(function ($count) {
+                $this->assertEquals(1, $count);
+            });
+
+        $this->channelManager
+            ->getConnectionsFromSet(0, Carbon::now()->subMinutes(2)->format('U'))
+            ->then(function ($expiredConnections) {
+                $this->assertCount(0, $expiredConnections);
+            });
+    }
+
+    public function test_events_are_processed_by_on_message_on_public_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $connection = $this->newActiveConnection(['public-channel']);
+
+        $message = new Mocks\Message([
+            'appId' => '1234',
+            'serverId' => 'different_server_id',
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'public-channel',
+                'test' => 'yes',
+            ],
+        ]);
+
+        $this->channelManager->onMessage(
+            $this->channelManager->getRedisKey('1234', 'public-channel'),
+            $message->getPayload()
+        );
+
+        // The message does not contain appId and serverId anymore.
+        $message = new Mocks\Message([
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'public-channel',
+                'test' => 'yes',
+            ],
+        ]);
+
+        $connection->assertSentEvent('some-event', $message->getPayloadAsArray());
+    }
+
+    public function test_events_get_replicated_across_connections_for_public_channels()
+    {
+        $this->runOnlyOnRedisReplication();
+
+        $connection = $this->newActiveConnection(['public-channel']);
+        $receiver = $this->newActiveConnection(['public-channel']);
+
+        $message = new Mocks\Message([
+            'appId' => '1234',
+            'serverId' => $this->channelManager->getServerId(),
+            'event' => 'some-event',
+            'data' => [
+                'channel' => 'public-channel',
+                'test' => 'yes',
+            ],
+            'socketId' => $connection->socketId,
+        ]);
+
+        $channel = $this->channelManager->find('1234', 'public-channel');
+
+        $channel->broadcastToEveryoneExcept(
+            $message->getPayloadAsObject(), $connection->socketId, '1234', true
+        );
+
+        $receiver->assertSentEvent('some-event', $message->getPayloadAsArray());
+
+        $this->getSubscribeClient()
+            ->assertNothingDispatched();
+
+        $this->getPublishClient()
+            ->assertCalledWithArgs('publish', [
+                $this->channelManager->getRedisKey('1234', 'public-channel'),
+                $message->getPayload(),
+            ]);
     }
 }
